// src/types/MembersType.ts
type Member = {
    id: number
    name: string
    address: string
    neighborhood: string
    zip_code: string
    cpf: string
    marital_status: string
    phone_number: number
    date_of_birth: number
    email: string
    social_network: string
    unit_served: string
    year_served: number
    number: string
    war_name: string
    obs: string
}

export type { Member }